+++ src/app/js/router.js
@@ -364,8 +364,10 @@ Y.Router = Y.extend(Router, Y.Base, {
     @param {Function|String} callback Callback function to call whenever this
         route is triggered. If specified as a string, the named function will be
         called on this router instance.
+
       @param {Object} callback.req Request object containing information about
           the request. It contains the following properties.
+
         @param {Array|Object} callback.req.params Captured parameters matched by
           the route path specification. If a string path was used and contained
           named parameters, then this will be a key/value hash mapping parameter
@@ -373,6 +375,8 @@ Y.Router = Y.extend(Router, Y.Base, {
           an array of subpattern matches starting at index 0 for the full match,
           then 1 for the first subpattern match, and so on.
         @param {String} callback.req.path The current URL path.
+        @param {Number} callback.req.pendingRoutes Number of matching routes
+          after this one in the dispatch chain.
         @param {Object} callback.req.query Query hash representing the URL query
           string, if any. Parameter names are keys, and are mapped to parameter
           values.
@@ -380,10 +384,12 @@ Y.Router = Y.extend(Router, Y.Base, {
         @param {String} callback.req.src What initiated the dispatch. In an
           HTML5 browser, when the back/forward buttons are used, this property
           will have a value of "popstate".
+
       @param {Object} callback.res Response object containing methods and
           information that relate to responding to a request. It contains the
           following properties.
         @param {Object} callback.res.req Reference to the request object.
+
       @param {Function} callback.next Callback to pass control to the next
         matching route. If you don't call this function, then no further route
         handlers will be executed, even if there are more that match. If you do
@@ -559,9 +565,9 @@ Y.Router = Y.extend(Router, Y.Base, {
                 } else {
                     req.params = matches.concat();
                 }
-                
+
                 req.pendingRoutes = routes.length;
-                
+
                 callback.call(self, req, res, req.next);
             }
         };